Add more IT and fix deletion bugs

+++ server/src/main/java/org/apache/iotdb/db/engine/memtable/AbstractMemTable.java
@@ -109,7 +109,7 @@ public abstract class AbstractMemTable implements IMemTable {
             deviceId,
             k -> {
               seriesNumber += schemaList.size();
-              totalPointsNumThreshold += avgSeriesPointNumThreshold * schemaList.size();
+              totalPointsNumThreshold += ((long) avgSeriesPointNumThreshold) * schemaList.size();
               return new AlignedWritableMemChunkGroup(schemaList);
             });
     for (IMeasurementSchema schema : schemaList) {
